requirements.txt
tqdm
numpy
xmltodict
h5py
opencv-python
argparse
matplotlib
scikit-learn
torchsummary
wandb
black
torchmetrics
